Medical-Bone-Age-main/backend/requirements.txt
fastapi
uvicorn
python-multipart
torch
torchvision
numpy
opencv-python-headless
ultralytics
matplotlib
requests
python-dotenv
